Change colors instead of material (WIP)
This is WIP as the feature is pretty basic, but it adds color information to briqs and lets players change it, supporting undo/redo and everything else that's nice.

The paint mode has been activated for this. The placer also takes color into account.

Code is fairly ugly I think, but it works.